my-actor/requirements.txt
apify-client>=1.1.0
beautifulsoup4>=4.12.0
httpx>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
prometheus-client>=0.17.0
gputil>=1.4.0
pillow>=10.0.0
opencv-python>=4.8.0
lxml>=4.9.0
